// src/app/core/models/GitUser.ts
export interface GitUser {
  login: String;
  id: number;
  node_id: String;
  avatar_url: String;
  gravatar_id?: String;
  url: String;
  html_url: String;
  followers_url: String;
  following_url: String;
  gists_url?: String;
  starred_url?: String;
  subscriptions_url?: String;
  organizations_url?: String;
  repos_url?: String;
  events_url?: String;
  received_events_url?: String;
  type: String;
  site_admin?: Boolean;
  name: String;
  company?: String;
  blog?: String;
  location?: String;
  email?: String;
  hireable?: Boolean;
  bio?: String;
  twitter_username?: String;
  public_repos: number;
  public_gists: number;
  followers: number;
  following: number;
  created_at: Date;
  updated_at: Date;
}